+++ cxx_mkspecs/clang_mkspecs.py
@@ -30,6 +30,19 @@ def cxx_apple_llvm70_x64(conf):
 
 
 @conf
+def cxx_apple_llvm73_x64(conf):
+    """
+    Detect and setup the 64-bit Apple LLVM 7.3 compiler
+    """
+    if conf.is_mkspec_platform('mac'):
+        conf.mkspec_clang_configure(7, 3)
+        conf.mkspec_add_common_flag('-m64')
+    else:
+        conf.fatal("This mkspec is not supported on {0}.".format(
+            conf.get_mkspec_platform()))
+
+
+@conf
 def cxx_clang35_x64(conf):
     """
     Detect and setup the clang 3.5 compiler for 64 bit
@@ -62,6 +75,24 @@ def cxx_clang36_x86(conf):
     Detect and setup the clang 3.6 compiler for 32 bit
     """
     conf.mkspec_clang_configure(3, 6)
+    conf.mkspec_add_common_flag('-m32')
+
+
+@conf
+def cxx_clang37_x64(conf):
+    """
+    Detect and setup the clang 3.7 compiler for 64 bit
+    """
+    conf.mkspec_clang_configure(3, 7)
+    conf.mkspec_add_common_flag('-m64')
+
+
+@conf
+def cxx_clang37_x86(conf):
+    """
+    Detect and setup the clang 3.7 compiler for 32 bit
+    """
+    conf.mkspec_clang_configure(3, 7)
     conf.mkspec_add_common_flag('-m32')